Add info_hash parameter in tracker request
Torrent trackers require a urlencoded info_hash parameter containing the SHA1
hash of the value of the info key of the torrent info file.

This should work for most trackers, but currently urlencodes unreserved
characters by adding percent chars. For maximum interoperability, this will
need to be changed so only non-ASCII characters are encoded.